backend/requirements.txt
allpairspy==2.5.1
annotated-types==0.6.0
approval_utilities==11.1.1
approvaltests==11.1.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
databricks-sdk==0.21.0
empty-files==0.0.9
exceptiongroup==1.2.0
execnet==2.0.2
Flask==3.0.2
Flask-Cors==4.0.0
google-auth==2.28.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mock==5.1.0
mrjob==0.7.4
packaging==24.0
pluggy==1.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyperclip==1.8.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
soupsieve==2.5
testfixtures==8.1.0
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
pyspark>=3.2.2
delta-spark>=2.0.0
